+++ core/collection.ts
@@ -16,7 +16,7 @@ export interface CollectionOptions {
   label?: string;
   description?: string;
   storage: Storage;
-  fields: Lume.CMS.ResolvedField;
+  fields?: Lume.CMS.ResolvedField;
   previewUrl?: PreviewUrl;
   views?: string[] | ((data?: Data) => string[] | undefined);
   documentName?: string | ((changes: Data) => string | undefined);
@@ -39,7 +39,7 @@ export default class Collection {
   label: string;
   description?: string;
   #storage: Storage;
-  #fields: Lume.CMS.ResolvedField;
+  #fields?: Lume.CMS.ResolvedField;
   previewUrl?: PreviewUrl;
   views?: string[] | ((data?: Data) => string[] | undefined);
   documentName?: string | ((changes: Data) => string | undefined);